✨ feat: Revised architecture and config
