+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["gas-benchmarks"]
+members = ["profiler"]
 resolver = "2"
 
-default-members = ["gas-benchmarks"]
+default-members = ["profiler"]
