--- clients/erigon/docker-compose.yml
@@ -1,30 +1,0 @@
-name: erigon
-services:
-  setup:
-    container_name: setup
-    user: "root:root"
-    image: erigontech/erigon:latest
-    volumes:
-      - ./data:/data
-      - ./genesis.json:/genesis.json
-    command: init --datadir=/data /genesis.json
-
-  node:
-    depends_on:
-      setup:
-        condition: service_completed_successfully
-    user: "root:root"
-    image: erigontech/erigon:latest
-    ports:
-      - "8551:8551"
-    volumes:
-      - ./data:/data
-      - ./jwt.hex:/config/jwt.hex
-    command:
-      - --nodiscover
-      - --datadir=/data
-      - --networkid=1700
-      - --authrpc.addr=0.0.0.0
-      - --authrpc.port=8551
-      - --authrpc.vhosts=*
-      - --authrpc.jwtsecret=/config/jwt.hex
