--- clients/nethermind/docker-compose.yml
@@ -1,51 +1,0 @@
-version: "3.9"
-services:
-  execution:
-    tty: true
-    environment:
-    - TERM=xterm-256color
-    - COLORTERM=truecolor
-    stop_grace_period: 30s
-    container_name: gas-execution-client
-    restart: unless-stopped
-    image: nethermindeth/nethermind:latest
-    networks:
-    - gas
-    volumes:
-    - ./data:/nethermind/data
-    - ./jwt.hex:/tmp/jwt/jwtsecret
-    - ./chainspec.json:/tmp/chainspec/chainspec.json
-    ports:
-    - "30304:30304/tcp"
-    - "30304:30304/udp"
-    - "8009:8009"
-    - "8545:8545"
-    - "8551:8551"
-    expose:
-    - 8545
-    - 8551
-    command:
-    - --config=holesky
-    - --Init.ChainSpecPath=/tmp/chainspec/chainspec.json
-    - --datadir=/nethermind/data
-    - --log=INFO
-    - --JsonRpc.Enabled=true
-    - --JsonRpc.Host=0.0.0.0
-    - --JsonRpc.Port=8545
-    - --JsonRpc.JwtSecretFile=/tmp/jwt/jwtsecret
-    - --JsonRpc.EngineHost=0.0.0.0
-    - --JsonRpc.EnginePort=8551
-    - --Network.DiscoveryPort=30304
-    - --HealthChecks.Enabled=true
-    - --Metrics.Enabled=true
-    - --Metrics.ExposePort=8009
-    - --Init.GenesisHash=0x9cbea0de83b440f4462c8280a4b0b4590cdb452069757e2c510cb3456b6c98cc
-    # - --Sync.MaxAttemptsToUpdatePivot=0
-    logging:
-      driver: json-file
-      options:
-        max-size: 10m
-        max-file: "10"
-networks:
-  gas:
-    name: gas-network
