--- clients/nethermind/jwt.hex
@@ -1,1 +1,0 @@
-0ed2c92023941418bd978b7f0daa545840d21f404e91173e62e6ef0bc41c91cb
