+++ profiler/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gas-benchmarks"
+name = "profiler"
 version = "0.1.0"
 edition = "2021"
